drop xxd dependency
This version will run on onecompiler and should still be posixly correct. (ie bourne shell not bash)

Instead of taking the first n bytes out of urandom and converting them to hex (using xxd) the trick here is to take bytes dropping any that are not in the range of allowed hex characters. This still ends up with the right number of hex digits, tho it's not particularly efficient.